add more unit test
