+++ internal/otelcollector/config/metric/gateway/processors.go
@@ -79,19 +79,6 @@ func makeDropIfInputSourceOtlpConfig() *FilterProcessor {
 	}
 }
 
-func makeDropDiagnosticMetricsForInput(inputSourceCondition string) *FilterProcessor {
-	var filterExpressions []string
-	metricNameConditions := createNameConditions([]string{"up", "scrape_duration_seconds", "scrape_samples_scraped", "scrape_samples_post_metric_relabeling", "scrape_series_added"})
-	excludeScrapeMetricsExpr := ottlexpr.JoinWithAnd(inputSourceCondition, ottlexpr.JoinWithOr(metricNameConditions...))
-	filterExpressions = append(filterExpressions, excludeScrapeMetricsExpr)
-
-	return &FilterProcessor{
-		Metrics: FilterProcessorMetrics{
-			Metric: filterExpressions,
-		},
-	}
-}
-
 func makeResolveServiceNameConfig() *TransformProcessor {
 	return &TransformProcessor{
 		ErrorMode:        "ignore",
@@ -143,14 +130,6 @@ func createNamespacesConditions(namespaces []string) []string {
 		namespacesConditions = append(namespacesConditions, ottlexpr.NamespaceEquals(ns))
 	}
 	return namespacesConditions
-}
-
-func createNameConditions(names []string) []string {
-	var nameConditions []string
-	for _, name := range names {
-		nameConditions = append(nameConditions, ottlexpr.NameAttributeEquals(name))
-	}
-	return nameConditions
 }
 
 func inputSourceEquals(inputSourceType metric.InputSourceType) string {
